提交代码

+++ aspnet-core/src/Finance.Application/PriceEval/Dto/CreateModelCountDto.cs
@@ -83,7 +83,7 @@ namespace Finance.PriceEval.Dto
         /// </summary>
         [Required]
         [Range(1, long.MaxValue)]
-        public virtual decimal Quantity { get; set; }
+        public virtual int Quantity { get; set; }
 
     }
 }
