+++ aspnet-core/src/Finance.Application/PriceEval/Dto/PcsListDto.cs
@@ -42,6 +42,6 @@ namespace Finance.PriceEval.Dto
         /// <summary>
         /// 数量
         /// </summary>
-        public virtual decimal Quantity { get; set; }
+        public virtual int Quantity { get; set; }
     }
 }
