+++ aspnet-core/src/Finance.Application/PriceEval/Dto/ShareCountInput.cs
@@ -17,6 +17,6 @@ namespace Finance.PriceEval.Dto
         /// <summary>
         /// 分摊数量
         /// </summary>
-        public virtual decimal Count { get; set; }
+        public virtual int Count { get; set; }
     }
 }
